--- web/src/components/landingpage/onboardingHost copy.css
@@ -1,232 +1,0 @@
-.formContainer {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  margin: auto;
-
-}
-
-.formHolder {
-  width: 100%;
-}
-
-.form-section,
-.map-section {
-  min-width: 45%;
-}
-
-.front-section {
-  width: 50%;
-}
-
-.formContainer label {
-  gap: 15px;
-}
-
-.mBottom {
-  margin-bottom: 50px;
-}
-
-.mTop {
-  margin-top: 15px;
-}
-
-.room-type label,
-.room-features label {
-  cursor: pointer;
-}
-
-.room-type {
-  display: flex;
-  flex-direction: column;
-  margin-right: 100px;
-}
-
-.quantity {
-  display: flex;
-  flex-direction: column;
-  max-width: 600px;
-  margin: auto;
-}
-
-.input-group {
-  margin-bottom: 5px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.input-group label {
-  width: 70%;
-  margin-right: 20px;
-  text-align: right;
-}
-
-.input-group .textInput {
-  width: auto; 
-  max-width: 100px;
-  padding: 10px;
-  border: 1px solid var(--secondary-color);
-  border-radius: 5px;
-}
-
-.formRow {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  flex-wrap: wrap;
-}
-
-.reviewInfo {
-  text-align: left;
-  margin-bottom: 20px;
-}
-
-.textInput,
-.radioInput {
-  padding: 11px 15px;
-  border-radius: 5px;
-  border: 1px solid var(--secondary-color);
-}
-
-.radioInput {
-  width: 15px;
-  height: 20px;
-  cursor: pointer;
-}
-
-.locationText {
-  min-width: 173px;
-  width: 60%;
-}
-
-.configurations label{
-    min-width: 250px;
-    margin-left: auto;
-    margin-right: auto;
-    font-size: 18px;
-    margin-bottom: 20px; 
-    white-space: nowrap;
-}
-
-.priceSlider {
-  width: 100%;
-  cursor: ew-resize;
-}
-
-.onboardingSectionTitle {
-  font-size: 20px;
-  font-weight: 500;
-  color: var(--secondary-color)
-}
-
-.file-upload {
-  background-color: #003366;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  cursor: pointer;
-  border-radius: 5px;
-}
-
-.file-upload:hover {
-  background-color: #001f4d; 
-}
-
-.file-input {
-  position: absolute;
-  top: 0;
-  left: 0;
-  opacity: 0;
-}
-
-@media screen and (max-width: 1520px) {
-  .formContainer label {
-    display: flex;
-    flex-direction: row;
-    text-align: left;
-    align-items: flex-start;
-    margin-bottom: 20px;
-  }
-  .formContainer {
-    text-align: left;
-  }
-}
-
-.file-input-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.custom-file-upload {
-  background-color: #003366;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  cursor: pointer;
-  border-radius: 5px;
-}
-
-.custom-file-upload:hover {
-  background-color: #001f4d;
-}
-
-.flex-row {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  margin: auto;
-}
-
-.image-container {
-  position: relative;
-  margin: 10px;
-}
-
-.image {
-  width: 150px;
-  height: 150px;
-  object-fit: cover;
-}
-
-.delete-button {
-  position: absolute;
-  top: 5px;
-  right: 5px;
-  background-color: #ff0000;
-  color: white;
-  border: none;
-  padding: 5px 10px;
-  border-radius: 50%;
-  cursor: pointer;
-}
-
-.delete-button:hover {
-  background-color: #cc0000;
-}
-
-.placeholder {
-  width: 150px;
-  height: 150px;
-  background-color: #f0f0f0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@media (max-width: 425px) {
-  .quantity {
-    margin: 0;
-  }
-  .nextButtons {
-    width: 40%;
-    margin-top: 25px;
-  }
-  .flex-row {
-    flex-direction: column;
-    width: 100%;
-    margin: auto;
-  }
-}
